backend/requirements.txt
fastapi==0.104.1
breeze-connect==1.0.63
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.2.1
python-dotenv==1.0.0
python-multipart==0.0.6
# Utils
python-dateutil==2.8.2
pandas==2.1.1
numpy==1.26.0
apscheduler==3.10.4
# API Documentation
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4